requirement.txt
Click==7.0
Django==2.2.5
Flask==1.1.1
Flask-API==1.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
gunicorn==20.0.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
protobuf==3.9.2
PyMySQL==0.9.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
Werkzeug==0.16.0
